requirements.txt
# Core Streamlit dependencies
streamlit>=1.28.0
streamlit-mic-recorder>=0.0.5

# AI/ML dependencies
google-generativeai>=0.3.0
langdetect>=1.0.9
langid>=1.1.6

# Audio processing
numpy>=1.24.0
soundfile>=0.12.1
scipy>=1.10.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.0.0